better unsigned integer handling
git-svn-id: https://flazr.svn.sourceforge.net/svnroot/flazr/trunk/flazr@80 5dee199b-8608-4927-992f-dc734163a847

+++ src/main/java/com/flazr/io/f4v/Box.java
@@ -44,7 +44,7 @@ public class Box {
         if (boxSize == 1) { // extended
             final byte[] extBytes = in.readBytes(8);
             final BigInteger bigLen = new BigInteger(1, extBytes);
-            payloadSize = bigLen.intValue() - 16;
+            payloadSize = bigLen.longValue() - 16;
         } else if (boxSize == 0) { // decided by parent bound
             payloadSize = endPos - in.position();
         } else {
